+++ src/config.js
@@ -1,0 +1,11 @@
+const toggles = [
+  "closeModalByClickingOutside",
+  "darkenModalOverlays",
+  "autoSave",
+  "noScrollBarOnTextAreas",
+  "showParentAsBreadcrumbs",
+  "moreProminentStateOnCards",
+  "copyNumberToClipboard",
+];
+
+export { toggles };